+++ src/main/java/org/springframework/samples/petclinic/security/model/ErrorResponseDto.java
@@ -2,11 +2,11 @@ package org.springframework.samples.petclinic.security.model;
 
 import org.springframework.http.HttpStatus;
 
-public class ErrorRes {
+public class ErrorResponseDto {
     HttpStatus httpStatus;
     String message;
 
-    public ErrorRes(HttpStatus httpStatus, String message) {
+    public ErrorResponseDto(HttpStatus httpStatus, String message) {
         this.httpStatus = httpStatus;
         this.message = message;
     }
